src/app/pages/products/product-list/product-list.component.ts
import { Component, OnDestroy, OnInit } from '@angular/core';
import { Product } from '@shared/models/product.model';
import { ProductsService } from '@shared/services/products.service';

@Component({
  selector: 'app-product-list',
  templateUrl: './product-list.component.html',
  styleUrls: ['./product-list.component.scss'],
})
export class ProductListComponent implements OnInit {
  products: Product[] = [];
  constructor(private productsService: ProductsService) {}

  ngOnInit(): void {
    this.loadProducts();
  }

  loadProducts(): void {
    this.productsService.getProducts().subscribe({
      next: (products: Product[]) => {
        this.products = products;
      },
    });
  }
}
